Rename entity classes for more clarity. Break out the id field into a true base entity in case we want tables without audit fields
