+++ genie-core/src/main/java/com/netflix/genie/core/jpa/entities/IdEntity.java
@@ -1,0 +1,50 @@
+/*
+ *
+ *  Copyright 2017 Netflix, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ *
+ */
+package com.netflix.genie.core.jpa.entities;
+
+import com.netflix.genie.core.jpa.entities.projections.IdProjection;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+import javax.persistence.Column;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+import java.io.Serializable;
+
+/**
+ * Base class which only provides an ID. For use cases when we don't care about audit fields like
+ * created, updated, entity version etc.
+ *
+ * @author tgianos
+ * @since 3.3.0
+ */
+@Getter
+@ToString
+@EqualsAndHashCode(of = "id")
+@MappedSuperclass
+public class IdEntity implements IdProjection, Serializable {
+
+    private static final long serialVersionUID = 7526472297322776147L;
+
+    @Id
+    @GeneratedValue
+    @Column(name = "id", nullable = false, updatable = false)
+    private long id;
+}
