+++ genie-core/src/main/java/com/netflix/genie/core/jpa/entities/projections/JobExecutionProjection.java
@@ -26,7 +26,7 @@ import java.util.Optional;
  * @author tgianos
  * @since 3.3.0
  */
-public interface JobExecutionProjection extends BaseProjection, JobHostNameProjection {
+public interface JobExecutionProjection extends AuditProjection, JobHostNameProjection {
 
     /**
      * Get the unique identifier of this job execution.
